Fixed charset when standard is used
